store/put server, client funcs into different files
